WIP: Create index page. (#30)
* Cleanup, and create components.

* Fix styling.

* Fix file names.

* Delete unused css

* GitHub comments.
