+++ client/src/app/about/about.component.ts
@@ -1,0 +1,14 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'about',
+  templateUrl: 'app/about/about.component.html',
+  styleUrls: ['app/about/about.component.css'],
+  providers: [],
+  directives: [],
+  pipes: []
+})
+
+export class About {
+  constructor() { }
+}
