+++ client/src/app/app.html
@@ -9,10 +9,6 @@
 	<a [routerLink]=" ['/About'] ">
     About
   </a>
-  |
-	<a [routerLink]=" ['/RepoBrowser/RepoList', {org: 'angular'}] ">
-    Github Repos
-  </a>
 </nav>
 
 <main>
