--- client/src/app/components/about/about.ts
@@ -1,22 +1,0 @@
-import {Component} from '@angular/core';
-import {Http} from '@angular/http';
-
-
-@Component({
-  selector: 'about',
-  templateUrl: 'app/components/about/about.html',
-  styleUrls: ['app/components/about/about.css'],
-  providers: [],
-  directives: [],
-  pipes: []
-})
-export class About {
-
-  constructor(http: Http) {
-
-  }
-
-  ngOnInit() {
-
-  }
-}
