--- client/src/app/components/home/home.ts
@@ -1,19 +1,0 @@
-import {Component} from '@angular/core';
-
-@Component({
-  selector: 'home',
-  templateUrl: 'app/components/home/home.html',
-  styleUrls: ['app/components/home/home.css'],
-  providers: [],
-  directives: [],
-  pipes: []
-})
-export class Home {
-
-  constructor() {}
-
-  ngOnInit() {
-
-  }
-
-}
