--- client/src/app/components/repo-browser/repo-browser.html
@@ -1,6 +1,0 @@
-<h3>GitHub Browser</h3>
-
-<input type="text" #repoName placeholder="Search Github Orgs" >
-<button (click)="searchForOrg(repoName.value)">Search Orgs</button>
-
-<router-outlet></router-outlet>
