--- client/src/app/components/repo-browser/repo-browser.ts
@@ -1,32 +1,0 @@
-import {Component} from '@angular/core';
-import {Router, RouteConfig, ROUTER_DIRECTIVES} from '@angular/router-deprecated';
-
-import {RepoList} from '../repo-list/repo-list';
-import {RepoDetail} from '../repo-detail/repo-detail';
-import {Github} from '../../services/github';
-
-@Component({
-  selector: 'repo-browser',
-  templateUrl: 'app/components/repo-browser/repo-browser.html',
-  styleUrls: ['app/components/repo-browser/repo-browser.css'],
-  providers: [ Github ],
-  directives: [ ROUTER_DIRECTIVES ],
-  pipes: []
-})
-@RouteConfig([
-  {path: '/:org',       component: RepoList,   name: 'RepoList'},
-  {path: '/:org/:name', component: RepoDetail, name: 'RepoDetail' },
-])
-export class RepoBrowser {
-
-  constructor(private router: Router, private github: Github) {}
-
-  searchForOrg(orgName: string) {
-    this.github.getOrg(orgName)
-      .subscribe(({name}) => {
-        console.log(name);
-        this.router.navigate(['RepoList', {org: orgName}]);
-      });
-  }
-
-}
