--- client/src/app/components/repo-detail/repo-detail.html
@@ -1,3 +1,0 @@
-<h2>{{ repoDetails.full_name }}</h2>
-
-<pre>this.repoDetails = {{ repoDetails | json }}</pre>
