--- client/src/app/components/repo-list/repo-list.html
@@ -1,8 +1,0 @@
-<h3>Repo list</h3>
-<ul>
-	<li *ngFor="let repo of repos | async">
-    <a [routerLink]=" ['RepoDetail', {org: repo.owner.login, name: repo.name}] ">
-      {{ repo.name }}
-    </a>
-  </li>
-</ul>
