+++ client/src/app/home/home.component.ts
@@ -1,0 +1,14 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'home',
+  templateUrl: 'app/home/home.component.html',
+  styleUrls: ['app/home/home.component.css'],
+  providers: [],
+  directives: [],
+  pipes: []
+})
+
+export class Home {
+  constructor() { }
+}
