+++ client/src/app/shared/search/search.component.html
@@ -1,0 +1,2 @@
+<h3>Search Component</h3>
+<p>This is the search component!</p>
