+++ client/src/app/shared/search/search.component.ts
@@ -1,0 +1,14 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'search',
+  templateUrl: 'app/search/search.component.html',
+  styleUrls: ['app/search/search.component.css'],
+  providers: [],
+  directives: [],
+  pipes: []
+})
+
+export class Search {
+  constructor() { }
+}
