+++ client/src/index.css
@@ -1,0 +1,4 @@
+body {
+    direction: rtl;  /* Right to Left */
+    background-color: #ebf4f8;
+}
