+++ client/src/index.html
@@ -30,12 +30,13 @@
   <!--<meta name="theme-color" content="#4e69a2">-->
   <!-- end favicon -->
 
+    <link href="index.css" rel="stylesheet">
 </head>
 <body>
 
-  <seed-app>
+  <app>
     Loading...
-  </seed-app>
+  </app>
 
   <script src="polyfills.bundle.js"></script>
   <script src="vendor.bundle.js"></script>
