+++ client/src/main.ts
@@ -1,0 +1,14 @@
+import { LocationStrategy, HashLocationStrategy } from '@angular/common';
+import { bootstrap } from '@angular/platform-browser-dynamic';
+import { provide, enableProdMode } from '@angular/core';
+import { HTTP_PROVIDERS } from '@angular/http';
+import { ROUTER_PROVIDERS } from '@angular/router-deprecated';
+
+import { App } from './app/app';
+
+bootstrap(App, [
+  HTTP_PROVIDERS,
+  ROUTER_PROVIDERS,
+  provide(LocationStrategy, {useClass: HashLocationStrategy})
+])
+.catch(err => console.error(err));
